+++ tests/e2e/signalling/ice_restart.rs
@@ -7,8 +7,8 @@ use function_name::named;
 use futures::{channel::mpsc::*, StreamExt as _};
 use medea::hashmap;
 use medea_client_api_proto::{
-    Command, Event, NegotiationRole, PeerConnectionState, PeerMetrics,
-    PeerUpdate, TrackId,
+    Command, Event, NegotiationRole, PeerConnectionState, PeerMetrics, TrackId,
+    TrackUpdate,
 };
 
 use crate::{
@@ -154,7 +154,7 @@ async fn ice_restart() {
     {
         let event = responder_rx.next().await.unwrap();
         match event {
-            Event::PeerUpdated {
+            Event::TracksApplied {
                 peer_id,
                 updates,
                 negotiation_role,
@@ -163,7 +163,7 @@ async fn ice_restart() {
                 assert_eq!(negotiation_role, Some(NegotiationRole::Offerer));
                 let is_ice_restart = updates
                     .iter()
-                    .find(|upd| matches!(upd, PeerUpdate::IceRestart))
+                    .find(|upd| matches!(upd, TrackUpdate::IceRestart))
                     .is_some();
                 assert!(is_ice_restart);
             }
@@ -190,7 +190,7 @@ async fn ice_restart() {
     {
         let event = publisher_rx.next().await.unwrap();
         match event {
-            Event::PeerUpdated {
+            Event::TracksApplied {
                 peer_id,
                 updates: _,
                 negotiation_role,