+++ tests/e2e/signalling/mod.rs
@@ -27,7 +27,7 @@ use awc::{
 use futures::{executor, stream::SplitSink, SinkExt as _, StreamExt as _};
 use medea_client_api_proto::{
     ClientMsg, Command, Event, IceCandidate, NegotiationRole, PeerId,
-    PeerUpdate, RpcSettings, ServerMsg, Track, TrackId,
+    RpcSettings, ServerMsg, Track, TrackId, TrackUpdate,
 };
 
 pub type MessageHandler =
@@ -315,7 +315,7 @@ impl StreamHandler<Result<Frame, WsProtocolError>> for TestMember {
                                     },
                                 });
                             }
-                            Event::PeerUpdated {
+                            Event::TracksApplied {
                                 peer_id,
                                 negotiation_role,
                                 updates,
@@ -323,7 +323,7 @@ impl StreamHandler<Result<Frame, WsProtocolError>> for TestMember {
                                 assert!(self.known_peers.contains(peer_id));
                                 updates.iter().for_each(|t| {
                                     use medea_client_api_proto::Direction;
-                                    if let PeerUpdate::Added(track) = t {
+                                    if let TrackUpdate::Added(track) = t {
                                         let mid = match &track.direction {
                                             Direction::Send { mid, .. }
                                             | Direction::Recv { mid, .. } => {