+++ proto/client-api/CHANGELOG.md
@@ -53,9 +53,9 @@ All user visible changes to this project will be documented in this file. This p
     - `RtcIceServerStats`.
 - `Cancelled` state to the `KnownIceCandidatePairState` ([#102]);
 - `is_required` field to `AudioSettings` and `VideoSettings` ([#106]);
-- `PeerUpdated` event with `PeerUpdate::Updated` and `PeerUpdate::Added` variants ([#81], [#105]);
+- `TracksApplied` event with `TrackUpdate::Updated` and `TrackUpdate::Added` variants ([#81], [#105]);
 - `ConnectionQualityUpdated` event ([#132]);
-- `IceRestart` variant to `PeerUpdate` ([#138]).
+- `IceRestart` variant to `TrackUpdate` ([#138]).
 
 [#28]: /../../pull/28
 [#58]: /../../pull/58